use CL2 k8s client
