+++ scenarios/perf-eval/storage-attach-detach-1000/terraform-inputs/aws.tfvars
@@ -60,15 +60,6 @@ eks_config_list = [{
   vpc_name    = "client-vpc"
   policy_arns = ["AmazonEKSClusterPolicy", "AmazonEKSVPCResourceController", "AmazonEKSWorkerNodePolicy", "AmazonEKS_CNI_Policy", "AmazonEC2ContainerRegistryReadOnly"]
   eks_managed_node_groups = [
-    # {
-    #   name           = "default"
-    #   ami_type       = "AL2_x86_64"
-    #   instance_types = ["m4.large"]
-    #   min_size       = 3
-    #   max_size       = 3
-    #   desired_size   = 3
-    #   capacity_type  = "ON_DEMAND"
-    # },
     {
       name           = "user"
       ami_type       = "AL2_x86_64"
@@ -80,9 +71,6 @@ eks_config_list = [{
     }
   ]
   eks_addons = [
-    # {
-    #   name            = "vpc-cni"
-    # },
     {
       name            = "aws-ebs-csi-driver"
       service_account = "ebs-csi-controller-sa"
